sentiment_bot/requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
langchain==0.2.16
google-generativeai==0.8.3
# nltk==3.9.1
lxml==5.3.0
langgraph==0.2.14
# spacy==3.7.6
# en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
# zh-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/zh_core_web_sm-3.7.0/zh_core_web_sm-3.7.0-py3-none-any.whl
# zh-core-web-md @ https://github.com/explosion/spacy-models/releases/download/zh_core_web_md-3.7.0/zh_core_web_md-3.7.0-py3-none-any.whl
# zh-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/zh_core_web_lg-3.7.0/zh_core_web_lg-3.7.0-py3-none-any.whl
# transformers==4.39.3
# torch==2.2.2
# flair==0.13.1